==> embody_scripts/step3/4_step3_data_concat.ipynb <==
## Concatenates data files from step3 to enter into master csv file, step3_compile/

        1. step3_1_3cat_conditionStats.csv
        2. step3_4_3cat_aggregate_stats.csv
        3. step3_2_distraction_fluctuation.csv

import os
import pandas as pd

# set variables
root_dir = '/Path/to/EMBODY/Files'
analysis = 'phase1_demo'

# step3 files to concatenate
files = ['step3_1_3cat_conditionStats.csv', \
         'step3_4_3cat_aggregate_stats.csv', \
         'step3_2_distraction_fluctuation.csv']

## Compile step3 stats across subjects - for SPSS

# Append all step3 data
appended_data = []

for f in files:
    
    os.chdir('%s/%s/results/step3_compile' % (root_dir, analysis))
    data = pd.read_csv('%s' % f, sep=',', index_col=0)
    # store DataFrame in list
    appended_data.append(data)
    
# Concatenate step3 meditation stats
df = pd.concat(appended_data, axis=1) # concat by columns
# remove duplicate "group" columns
df = df.loc[:, ~df.columns.duplicated()]
    
# write DataFrame to csv
os.chdir('%s/%s/results/step3_compile/' % (root_dir, analysis))
df.to_csv('step3_master.csv')
